requirements.txt
Django==1.9
oauthlib==1.0.3
Pillow==3.0.0
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.13
requests==2.9.0
requests-oauthlib==0.6.0
six==1.10.0
Unidecode==0.4.18
wheel==0.24.0
whitenoise==2.0.6
